+++ .github/workflows/pipeline.yml
@@ -18,7 +18,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r appMedicos/project/requirements.txt
+        pip install -r appPacientes/project/requirements.txt
 
           
 
@@ -36,7 +36,7 @@ jobs:
             distribution: 'adopt'
             package: '*.jar'
         - name: Publish package
-          run: pip install -r appMedicos/project/requirements.txt
+          run: pip install -r appPacientes/project/requirements.txt
           env:
             GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}		  
 
@@ -52,6 +52,6 @@ jobs:
           pip install behave
     - name: Teste
       run:
-          #COMANDO OK > behave appMedicos/project/features/medicos.feature
+          #COMANDO OK > behave appPacientes/project/features/pacientes.feature
           python -m pip install --upgrade pip
-          pip debug -r appMedicos/project/requirements.txt
+          pip debug -r appPacientes/project/requirements.txt
