--- appMedicos/docker-compose.yml
@@ -1,32 +1,0 @@
-services:
-
-  medicosapp:
-    container_name: medicos
-    build:
-      context: .
-      dockerfile: project/Dockerfile
-    ports:
-      - "8004:8004"
-    environment:
-      - PG_USER=postgres
-      - PG_PASSWORD=postgres
-      - PG_DB=postgres
-      - PG_PORT=5432
-      - PG_HOST=dbMedicos
-    depends_on:
-      - dbMedicos
-
-  dbMedicos:
-    container_name: dbMedicos
-    image: postgres:12
-    environment:
-      - POSTGRES_DB=postgres
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=postgres
-    ports:
-      - "5435:5432"
-    volumes:
-      - pgdata:/var/lib/postgresql/data
-
-volumes:
-  pgdata: {}
