--- appMedicos/documentos/hk-medicos.postman_collection.json
@@ -1,198 +1,0 @@
-{
-	"info": {
-		"_postman_id": "743330cb-7446-4af5-884d-88ff0c3b8c9e",
-		"name": "hk-medicos",
-		"description": "Projeto criado para atender a demanda de uma operadora de saúde através de um sistema de agendamento de consultas.",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-		"_exporter_id": "33097875",
-		"_collection_link": "https://winter-desert-869279.postman.co/workspace/bf09e259-9dfb-40ec-823b-a85d2a12f07a/collection/33097875-743330cb-7446-4af5-884d-88ff0c3b8c9e?action=share&source=collection_link&creator=33097875"
-	},
-	"item": [
-		{
-			"name": "Consultar medicos",
-			"event": [
-				{
-					"listen": "test",
-					"script": {
-						"exec": [
-							""
-						],
-						"type": "text/javascript",
-						"packages": {}
-					}
-				}
-			],
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "localhost:8004/medicos",
-					"host": [
-						"localhost"
-					],
-					"port": "8004",
-					"path": [
-						"medicos"
-					]
-				},
-				"description": "API para consultar medicos cadastrados, não é necessário informar nada no body."
-			},
-			"response": []
-		},
-		{
-			"name": "Criar medico",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"name\": \"Primeiro Doutor\",\r\n    \"cpf\": \"12345678910\",\r\n    \"crm\": \"112233\",\r\n    \"email\": \"ba@ba.com.br\",\r\n    \"senha\": \"1!med@system\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "localhost:8004/users/create",
-					"host": [
-						"localhost"
-					],
-					"port": "8004",
-					"path": [
-						"users",
-						"create"
-					]
-				},
-				"description": "API para cadastrar novos médicos, necessário informar os 5 dados abaixo:\n\n- nome\n    \n- cpf\n    \n- crm\n    \n- email\n    \n- senha"
-			},
-			"response": []
-		},
-		{
-			"name": "Atualizar medicos",
-			"request": {
-				"method": "PUT",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"id\": \"2\",\r\n    \"name\": \"Primeiro Doutor\",\r\n    \"cpf\": \"12345678910\",\r\n    \"crm\": \"112233\",\r\n    \"email\": \"ba@ba.com.br\",\r\n    \"senha\": \"1!med@system\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "localhost:8004/users/update/1",
-					"host": [
-						"localhost"
-					],
-					"port": "8004",
-					"path": [
-						"users",
-						"update",
-						"1"
-					]
-				},
-				"description": "API para atualizar médicos cadastrados, necessário informar os 6 dados abaixo:\n\n- id\n    \n- nome\n    \n- cpf\n    \n- crm\n    \n- email\n    \n- senha"
-			},
-			"response": []
-		},
-		{
-			"name": "Deletar medico",
-			"request": {
-				"method": "DELETE",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "localhost:8004/users/delete/1",
-					"host": [
-						"localhost"
-					],
-					"port": "8004",
-					"path": [
-						"users",
-						"delete",
-						"1"
-					]
-				},
-				"description": "API para deletar um médico cadastrado, necessário informar o id do usuário no endpoint."
-			},
-			"response": []
-		},
-		{
-			"name": "Consultar medico pelo ID",
-			"protocolProfileBehavior": {
-				"disableBodyPruning": true
-			},
-			"request": {
-				"method": "GET",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "localhost:8004/users/read/1",
-					"host": [
-						"localhost"
-					],
-					"port": "8004",
-					"path": [
-						"users",
-						"read",
-						"1"
-					]
-				},
-				"description": "API para consultar médicos cadastrados, filtrando pelo id. Necessário informar o id no final do endpoint."
-			},
-			"response": []
-		},
-		{
-			"name": "Consultar medico pelo CPF",
-			"protocolProfileBehavior": {
-				"disableBodyPruning": true
-			},
-			"request": {
-				"method": "GET",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "localhost:8004/users/readcpf/12345678910",
-					"host": [
-						"localhost"
-					],
-					"port": "8004",
-					"path": [
-						"users",
-						"readcpf",
-						"12345678910"
-					]
-				},
-				"description": "API para consultar médicos cadastrados, filtrando pelo CPF. Necessário informar o CPF no final do endpoint."
-			},
-			"response": []
-		}
-	]
-}
