--- appMedicos/kubernetes/db_deployment.yml
@@ -1,67 +1,0 @@
-###########################
-# Deployment
-###########################
-
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: postgres-deployment
-spec:
-  replicas: 2
-  selector:
-    matchLabels:
-      app: postgres-container
-  template:
-    metadata:
-      labels:
-        app: postgres-container
-        tier: backend
-    spec:
-      containers:
-        - name: db
-          image: postgres:12
-          env:
-            - name: POSTGRES_HOST_AUTH_METHOD
-              value: trust
-#            - name: POSTGRES_USER
-#              valueFrom:
-#                secretKeyRef:
-#                  name: postgres-credentials
-#                  key: POSTGRES_USER
-#            - name: POSTGRES_PASSWORD
-#              valueFrom:
-#                secretKeyRef:
-#                  name: postgres-credentials
-#                  key: POSTGRES_PASSWORD
-#            - name: POSTGRES_DB
-#              valueFrom:
-#                configMapKeyRef:
-#                  name: appMedicos-variables
-#                  key: POSTGRES_DB
-          ports:
-            - containerPort: 5432
-          volumeMounts:
-            - name: postgres-volume-mount
-              mountPath: /var/lib/postgresql/data
-      volumes:
-        - name: postgres-volume-mount
-          emptyDir: {}
----
-###########################
-# Persistent Volume Claim
-###########################
-
-kind: PersistentVolumeClaim
-apiVersion: v1
-metadata:
-  name: postgres-pvc
-  labels:
-     type: local
-spec:
-  #storageClassName: manual
-  accessModes:
-    - ReadWriteOnce
-  resources:
-    requests:
-      storage: 200M
- # volumeName: postgres-pv
