--- appMedicos/kubernetes/secret.yml
@@ -1,8 +1,0 @@
-apiVersion: v1
-kind: Secret
-metadata:
-  name: appMedicos-credentials
-type: Opaque
-data:
-  POSTGRES_USER: postgres
-  POSTGRES_PASSWORD: postgres
