--- appMedicos/project/behave.ini
@@ -1,3 +1,0 @@
-[behave]
-default_path = features
-paths = features/medicos.feature
