--- appMedicos/project/django.sh
@@ -1,11 +1,0 @@
-#!/bin/bash
-echo "Creating Migrations..."
-python manage.py makemigrations medicos
-echo ====================================
-
-echo "Starting Migrations..."
-python manage.py migrate
-echo ====================================
-
-echo "Starting Server..."
-python manage.py runserver 0.0.0.0:8004
