--- appMedicos/project/infrastructure/config/medicoapps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-from entities import usuarios
-
-class DjangoappConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'medicos'
