+++ appPacientes/docker-compose.yml
@@ -1,0 +1,32 @@
+services:
+
+  pacientesapp:
+    container_name: pacientes
+    build:
+      context: .
+      dockerfile: project/Dockerfile
+    ports:
+      - "8003:8003"
+    environment:
+      - PG_USER=postgres
+      - PG_PASSWORD=postgres
+      - PG_DB=postgres
+      - PG_PORT=5432
+      - PG_HOST=dbPacientes
+    depends_on:
+      - dbPacientes
+
+  dbPacientes:
+    container_name: dbPacientes
+    image: postgres:12
+    environment:
+      - POSTGRES_DB=postgres
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=postgres
+    ports:
+      - "5434:5432"
+    volumes:
+      - pgdata:/var/lib/postgresql/data
+
+volumes:
+  pgdata: {}
