+++ appPacientes/documentos/hk-pacientes.postman_collection.json
@@ -1,0 +1,198 @@
+{
+	"info": {
+		"_postman_id": "3a7e7192-90ad-4637-bdb9-cf77a0c224a4",
+		"name": "hk-pacientes",
+		"description": "Projeto criado para atender a demanda de uma operadora de saúde através de um sistema de agendamento de consultas.",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
+		"_exporter_id": "33097875",
+		"_collection_link": "https://winter-desert-869279.postman.co/workspace/bf09e259-9dfb-40ec-823b-a85d2a12f07a/collection/33097875-3a7e7192-90ad-4637-bdb9-cf77a0c224a4?action=share&source=collection_link&creator=33097875"
+	},
+	"item": [
+		{
+			"name": "Consultar pacientes",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							""
+						],
+						"type": "text/javascript",
+						"packages": {}
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "localhost:8004/pacientes",
+					"host": [
+						"localhost"
+					],
+					"port": "8004",
+					"path": [
+						"pacientes"
+					]
+				},
+				"description": "API para consultar pacientes cadastrados, não é necessário informar nada no body."
+			},
+			"response": []
+		},
+		{
+			"name": "Criar paciente",
+			"request": {
+				"method": "POST",
+				"header": [],
+				"body": {
+					"mode": "raw",
+					"raw": "{\r\n    \"name\": \"Primeiro Paciente\",\r\n    \"cpf\": \"06743138010\",\r\n    \"email\": \"ba@ba.com.br\",\r\n    \"senha\": \"1!paci@system\"\r\n}",
+					"options": {
+						"raw": {
+							"language": "json"
+						}
+					}
+				},
+				"url": {
+					"raw": "localhost:8004/pacientes/create",
+					"host": [
+						"localhost"
+					],
+					"port": "8004",
+					"path": [
+						"pacientes",
+						"create"
+					]
+				},
+				"description": "API para cadastrar novos pacientes, necessário informar os 4 dados abaixo:\n\n- nome\n    \n- cpf\n    \n- email\n    \n- senha"
+			},
+			"response": []
+		},
+		{
+			"name": "Atualizar paciente",
+			"request": {
+				"method": "PUT",
+				"header": [],
+				"body": {
+					"mode": "raw",
+					"raw": "{\r\n    \"id\": \"2\",\r\n    \"name\": \"Primeiro Paciente\",\r\n    \"cpf\": \"06743138010\",\r\n    \"email\": \"ba@ba.com.br\",\r\n    \"senha\": \"1!paci@system\"\r\n}",
+					"options": {
+						"raw": {
+							"language": "json"
+						}
+					}
+				},
+				"url": {
+					"raw": "localhost:8004/pacientes/update/1",
+					"host": [
+						"localhost"
+					],
+					"port": "8004",
+					"path": [
+						"pacientes",
+						"update",
+						"1"
+					]
+				},
+				"description": "API para atualizar pacientes cadastrados, necessário informar os 6 dados abaixo:\n\n- id\n    \n- nome\n    \n- cpf\n    \n- email\n    \n- senha"
+			},
+			"response": []
+		},
+		{
+			"name": "Deletar paciente",
+			"request": {
+				"method": "DELETE",
+				"header": [],
+				"body": {
+					"mode": "raw",
+					"raw": "",
+					"options": {
+						"raw": {
+							"language": "json"
+						}
+					}
+				},
+				"url": {
+					"raw": "localhost:8004/pacientes/delete/1",
+					"host": [
+						"localhost"
+					],
+					"port": "8004",
+					"path": [
+						"pacientes",
+						"delete",
+						"1"
+					]
+				},
+				"description": "API para deletar um paciente cadastrado, necessário informar o id do usuário no endpoint."
+			},
+			"response": []
+		},
+		{
+			"name": "Consultar paciente pelo ID",
+			"protocolProfileBehavior": {
+				"disableBodyPruning": true
+			},
+			"request": {
+				"method": "GET",
+				"header": [],
+				"body": {
+					"mode": "raw",
+					"raw": "",
+					"options": {
+						"raw": {
+							"language": "json"
+						}
+					}
+				},
+				"url": {
+					"raw": "localhost:8004/pacientes/read/1",
+					"host": [
+						"localhost"
+					],
+					"port": "8004",
+					"path": [
+						"pacientes",
+						"read",
+						"1"
+					]
+				},
+				"description": "API para consultar pacientes cadastrados, filtrando pelo id. Necessário informar o id no final do endpoint."
+			},
+			"response": []
+		},
+		{
+			"name": "Consultar paciente pelo CPF",
+			"protocolProfileBehavior": {
+				"disableBodyPruning": true
+			},
+			"request": {
+				"method": "GET",
+				"header": [],
+				"body": {
+					"mode": "raw",
+					"raw": "",
+					"options": {
+						"raw": {
+							"language": "json"
+						}
+					}
+				},
+				"url": {
+					"raw": "localhost:8004/pacientes/readcpf/06743138010",
+					"host": [
+						"localhost"
+					],
+					"port": "8004",
+					"path": [
+						"pacientes",
+						"readcpf",
+						"06743138010"
+					]
+				},
+				"description": "API para consultar pacientes cadastrados, filtrando pelo CPF. Necessário informar o CPF no final do endpoint."
+			},
+			"response": []
+		}
+	]
+}
