+++ appPacientes/kubernetes/app_deployment.yml
@@ -1,0 +1,65 @@
+###########################
+# Deployment
+###########################
+
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: django-postgre-deploy
+  labels:
+    app: django-postgre
+spec:
+  replicas: 2
+  selector:
+    matchLabels:
+      app: django-postgre
+  template:
+    metadata:
+      labels:
+        app: django-postgre
+    spec:
+      containers:
+      - name: django-appPacientes
+        image: appPacientes-djangoapp:latest
+        imagePullPolicy: Never
+        command: ["bash", "-e", "./django.sh"]
+        ports:
+        - containerPort: 8000
+        env:
+        - name: POSTGRES_HOST_AUTH_METHOD
+          value: trust
+#        - name: POSTGRES_USER
+#          valueFrom:
+#            secretKeyRef:
+#              name: postgres-credentials
+#              key: POSTGRES_USER
+#        - name: POSTGRES_PASSWORD
+#          valueFrom:
+#            secretKeyRef:
+#              name: postgres-credentials
+#              key: POSTGRES_PASSWORD
+#        - name: PG_DB
+#          valueFrom:
+#            configMapKeyRef:
+#              name: appPacientes-variables
+#              key: PG_DB
+        - name: PG_HOST
+          valueFrom:
+            configMapKeyRef:
+              name: appPacientes-variables
+              key: PG_HOST
+        - name: DB_PORT
+          valueFrom:
+            configMapKeyRef:
+              name: appPacientes-variables
+              key: DB_PORT
+        resources:
+          requests:
+            cpu: 100m
+#        volumeMounts:
+#        - name: postgres-volume-mount
+#          mountPath: /var/lib/postgresql/data
+#      volumes:
+#        - name: postgres-pv
+#          persistentVolumeClaim:
+#            claimName: postgres-pvc
