+++ appPacientes/kubernetes/config_map.yml
@@ -1,0 +1,20 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: appPacientes-variables
+data:
+  #env variables for the postgres component
+  #POSTGRES_USER: postgres
+  #POSTGRES_DB: appPacientes
+
+  #env variables for the backend component
+  DJANGO_ENV: 'development'
+  DEBUG: "1"
+  SECRET_KEY: 'django-insecure-e7im26dbj!(m-)^a3#z5&5d($lrk3h+&)ckuk*&gw$@pzp!g8v'
+
+ # DATABASE: db
+  DB_ENGINE: "django.db.backends.postgresql"
+  #PG_DB: postgres
+ # PG_USER: postgres
+  PG_HOST: db
+  DB_PORT: "5432"
