+++ appPacientes/kubernetes/db_service.yml
