+++ appPacientes/kubernetes/secret.yml
@@ -1,0 +1,8 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: appPacientes-credentials
+type: Opaque
+data:
+  POSTGRES_USER: postgres
+  POSTGRES_PASSWORD: postgres
