+++ appPacientes/project/behave.ini
@@ -1,0 +1,3 @@
+[behave]
+default_path = features
+paths = features/pacientes.feature
