+++ appPacientes/project/infrastructure/presenters/pacientesserializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from entities.pacientes.medicosmodels import Paciente
+
+class MedicoSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Paciente
+        fields = '__all__'
