+++ appPacientes/project/tests/stress.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+foi i in {1..1000}; do
+  curl localhost:31000/pacientes
+  sleep $1
+  done
